+++ Objective C/Utils/RakCenteredTextFieldCell.m
@@ -22,7 +22,7 @@
 	
 	// Center that in the proposed rect
 	
-	float heightDelta = newRect.size.height - textSize.height;
+	double heightDelta = newRect.size.height - textSize.height;
 	if (heightDelta > 0)
 	{
 		newRect.size.height -= heightDelta;
